Refactor WordPressKit->JetpackStats mapping in StatsService
